Module_5/Tasks44/DRFLearning/tutorial/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
pip==21.3.1
pytz==2021.3
setuptools==58.5.3
sqlparse==0.4.2
wheel==0.37.0
